+++ .github/workflows/pull-request.yml
@@ -11,7 +11,7 @@ jobs:
   unix:
     strategy:
       matrix:
-        version: [ 18, 20, 22 ]
+        version: [ 20, 22, 24 ]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
