--- index.js
@@ -1,16 +1,0 @@
-require('./lib/types.js');
-require('./lib/memory.js');
-require('./lib/hooks.js');
-require('./lib/actions.js');
-require('./lib/waits.js');
-require('./lib/validations.js');
-require('./lib/cookies.js');
-require('./lib/localSessionStorage.js');
-require('./lib/mobile.js');
-require('./lib/execute.js');
-require('./lib/mock.js');
-require('./lib/intercept.js');
-require('./lib/poDefine.js');
-require('./lib/mouseActions.js');
-require('./lib/keyboardActions.js');
-require('./lib/dialog.js');
