+++ package.json
@@ -1,9 +1,13 @@
 {
   "name": "@qavajs/steps-wdio",
-  "version": "2.8.0",
+  "version": "3.0.0",
   "description": "steps to interact with webdriverio",
-  "main": "./index.js",
-  "types": "./index.d.ts",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "exports": {
+    ".": "./lib/index.js",
+    "./steps": "./lib/steps.js"
+  },
   "scripts": {
     "build": "tsc",
     "test": "vitest run",
@@ -36,8 +40,8 @@
     "express": "^5.1.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.0",
-    "@types/express": "^5.0.2"
+    "vitest": "^3.2.3",
+    "@types/express": "^5.0.3"
   },
   "dependencies": {
     "webdriverio": "^9.15.0"
