+++ src/index.ts
@@ -1,0 +1,2 @@
+export type { IQavajsWdioConfig } from './IQavajsWdioConfig';
+export { locator } from './pageObject';
