+++ steps.js
@@ -1,0 +1,1 @@
+exports = require('lib/steps.js');
